+++ test/EasyPost/Fixture.php
@@ -280,4 +280,20 @@ class Fixture
             'email'     => 'test@example.com',
         ];
     }
+
+    public static function rma_form_options()
+    {
+        return [
+            'barcode' => 'RMA12345678900',
+            'line_items' => [
+                [
+                    'product' => [
+                        'title' => 'Square Reader',
+                        'barcode' => '855658003251',
+                    ],
+                    'units' => 8,
+                ],
+            ],
+        ];
+    }
 }